oddsAndEven created
Odds and Even game. It's an interactive game that asks the user(player) to choose either odd or even, they are asked to input a specific number of fingers, then the computer picks a random number, and finally, the sum of the numbers is calculated. If the sum is an even number and the user chose even, then the user wins, else the computer wins. Visa versa for odd.